Feature/vandana

+++ src/modules/pos/outlet/outlet.component.ts
@@ -1,9 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { map } from 'rxjs/operators';
+import { AppToastService } from '@modules/shared-module/services/app-toast.service';
 
 import { OutletDataService } from '../outlet-data.service';
-import { Data } from '../outletData';
 
 @Component({
   selector: 'sb-outlet',
@@ -12,7 +11,7 @@ import { Data } from '../outletData';
 })
 export class OutletComponent implements OnInit {
 
-  public odata: any;
+  public outletData: any;
   public length = 0;
   public total = 0;
   public id = 0;
@@ -25,7 +24,9 @@ export class OutletComponent implements OnInit {
 
   constructor(private outletService: OutletDataService,
     private router: Router,
-    private route: ActivatedRoute) { }
+    private route: ActivatedRoute,
+    private toast: AppToastService
+  ) { }
 
   ngOnInit(): void {
     this.getOutletData();
@@ -35,39 +36,48 @@ export class OutletComponent implements OnInit {
   getOutletData() {
     this.showloader = true
     this.outletService.getOutletData(this.page).subscribe(result => {
-      this.odata = result.outlets.data
-      this.length = result.outlets.per_page;
+      this.outletData = result.outlets.data
+      this.length = this.outletData.length;
       this.total = result.outlets.total;
       this.showloader = false
-    })
+    }, err => {
+      this.toast.error('Error', 'Server error.')
+      this.showloader = false
+    });
   }
 
   // For updating data on page change
   onPageChange(event: number) {
     this.page = event;
     this.getOutletData();
-    console.log('Here >>>', this.page, this.odata);
+    console.log('Here >>>', this.page, this.outletData);
   }
 
   // For deleting outlet
   deleteRow(id: number) {
-
-    this.outletService.deleteOutlet(id).subscribe(data => {
-      this.getOutletData();
-    });
-    console.log(this.odata);
+    if (confirm('Are you sure you want to delete?')) {
+      this.outletService.deleteOutlet(id).subscribe(data => {
+        this.getOutletData();
+        this.toast.success('Success', 'Deleted successfully.');
+      }, err => {
+        this.toast.error('Error', 'Server error.');
+      });
+    }
   }
 
   // For searching outlets table data
   search(event: any) {
     this.showloader = true
     this.outletService.searchOutlet(this.searchValue).subscribe(res => {
-      this.odata = res.outlets.data
-      this.length = this.odata.length;
+      this.outletData = res.outlets.data
+      this.length = this.outletData.length;
       this.total = res.outlets.total;
       this.showloader = false
-      console.log(this.odata.length)
-    })
+      console.log(this.outletData.length)
+    }, err => {
+      this.toast.error('Error', 'Server error.')
+      this.showloader = false
+    });
   }
 
 }
